requirements.txt
speech_recognition
pyttsx3
requests
openai
gtts
pygame
pyautogui
keyboard
subprocess
pygetwindow
screen_brightness_control
google-auth
google-auth-oauthlib
google-api-python-client
datetime
difflib
uuid
ctypes
threading
time
re